ajustes realizados nos verbos Delete de TemaRepository e PostagemRepository

+++ sts/blogpessoal/src/main/java/com/generation/blogpessoal/controller/PostagemController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.blogpessoal.model.Postagem;
 import com.generation.blogpessoal.repository.PostagemRepository;
@@ -23,74 +24,67 @@ import com.generation.blogpessoal.repository.TemaRepository;
 
 @RestController // Indica classe controladora do tipo rest
 @RequestMapping("/postagens") // Indica o caminho padrão dessa classe controladora
-@CrossOrigin(origins = "*", allowedHeaders = "*") // Configura a comunicação entre o Front e o Backend que podem estar rodando em servidores diferentes
+@CrossOrigin(origins = "*", allowedHeaders = "*") // Configura a comunicação entre o Front e o Backend que podem estar
+													// rodando em servidores diferentes
 public class PostagemController {
-	
+
 	@Autowired // Instancia o objeto automaticamente pelo Spring
-	private PostagemRepository postagemRepository; 
-	
+	private PostagemRepository postagemRepository;
+
 	@Autowired
-	private TemaRepository temaRepository; 
-	
+	private TemaRepository temaRepository;
+
 	@GetMapping
 	public ResponseEntity<List<Postagem>> getAll() { // Vai retornar uma lista do que está na classe postagem
-		
-	return	ResponseEntity.ok(postagemRepository.findAll()); // Encontra todos os objetos postagemRepository
-	}		
-	
-	@GetMapping ("/{id}") //  Dentro das {} há uma variável de caminho
-	public ResponseEntity<Postagem> getById(@PathVariable Long id){
-		
-		return postagemRepository.findById(id)
-				.map(resposta -> ResponseEntity.ok(resposta))
+
+		return ResponseEntity.ok(postagemRepository.findAll()); // Encontra todos os objetos postagemRepository
+	}
+
+	@GetMapping("/{id}") // Dentro das {} há uma variável de caminho
+	public ResponseEntity<Postagem> getById(@PathVariable Long id) {
+
+		return postagemRepository.findById(id).map(resposta -> ResponseEntity.ok(resposta))
 				.orElse(ResponseEntity.notFound().build());
 	}
-	
-	@GetMapping ("/titulo/{titulo}")
-	public ResponseEntity<List<Postagem>> getByTitulo(@PathVariable String titulo) { 
-		
-	return	ResponseEntity.ok(postagemRepository.findAllByTituloContainingIgnoreCase(titulo)); 
-	}	
-	
+
+	@GetMapping("/titulo/{titulo}")
+	public ResponseEntity<List<Postagem>> getByTitulo(@PathVariable String titulo) {
+
+		return ResponseEntity.ok(postagemRepository.findAllByTituloContainingIgnoreCase(titulo));
+	}
+
 	@PostMapping
-	public ResponseEntity<Postagem> postPostagem(@Valid @RequestBody Postagem postagem){
-		
-		if(temaRepository.existsById(postagem.getTema().getId()))
-			return ResponseEntity.status(HttpStatus.CREATED)
-			.body(postagemRepository.save(postagem));
+	public ResponseEntity<Postagem> postPostagem(@Valid @RequestBody Postagem postagem) {
+
+		if (temaRepository.existsById(postagem.getTema().getId()))
+			return ResponseEntity.status(HttpStatus.CREATED).body(postagemRepository.save(postagem));
 
 		return ResponseEntity.notFound().build();
+	}
+
+	@PutMapping
+	public ResponseEntity<Postagem> putPostagem(@Valid @RequestBody Postagem postagem) {
+		if (temaRepository.existsById(postagem.getTema().getId())) {
+			return postagemRepository.findById(postagem.getId())
+					.map(resposta -> ResponseEntity.status(HttpStatus.OK).body(postagemRepository.save(postagem)))
+					.orElse(ResponseEntity.notFound().build());
+
 		}
-	
-	
-	@PutMapping 
-	    public ResponseEntity<Postagem> putPostagem(@Valid @RequestBody Postagem postagem) {
-		if(temaRepository.existsById(postagem.getTema().getId())) {
-			return postagemRepository.findById(postagem.getId()) 
-					.map(resposta -> ResponseEntity.status(HttpStatus.OK)
-					.body(postagemRepository.save(postagem))) 
-					.orElse(ResponseEntity.notFound().build()); 
-					
-		} 
-			return ResponseEntity.notFound().build();
-	    }
-								
-	
+		return ResponseEntity.notFound().build();
+	}
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<?> deletePostagem(@PathVariable Long id) {
 		
-		return postagemRepository.findById(id)
-				.map(resposta -> {
+		if (postagemRepository.existsById(id)){
+		   return postagemRepository.findById(id).map(resposta -> {
 					postagemRepository.deleteById(id);
 					return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 				})
 				.orElse(ResponseEntity.notFound().build());
+		}
+		throw new ResponseStatusException(HttpStatus.NOT_FOUND, "O Id inserido não corresponde a nenhuma postagem!");
 	}
-
-
+	
 
 }
-
-
-
